requirements.txt
boto==2.49.0
boto3==1.9.47
botocore==1.12.47
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
gensim==3.6.0
gspread==3.0.1
httplib2==0.12.0
idna==2.7
jmespath==0.9.3
mecab-python3==0.7
numpy==1.15.4
oauth2client==4.1.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.5
requests==2.20.1
rsa==4.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
urllib3==1.24.1
